requirements.txt
absl-py==0.14.0
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.1
attrs==20.3.0
autopep8==1.5.4
awscli==1.18.69
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bleach==3.2.3
blinker==1.4
boto3==1.9.253
botocore==1.16.19
Brlapi==0.7.0
CacheControl==0.12.6
cachetools==4.2.3
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
charset-normalizer==2.0.7
chrome-gnome-shell==0.0.0
clang==5.0
Click==7.0
colorama==0.4.3
coloredlogs==7.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cwltool==2.0.20200224214940
cycler==0.10.0
datasets==1.15.1
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
dill==0.3.4
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
docopt==0.6.2
docutils==0.16
duplicity==0.8.12.0
elasticsearch==7.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
flatbuffers==1.12
frozenlist==1.2.0
fsspec==2021.11.0
future==0.18.2
gast==0.4.0
gensim==4.1.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
graphviz==0.17
grpcio==1.41.0
h5py==3.1.0
html5lib==1.0.1
httplib2==0.14.0
huggingface-hub==0.1.2
humanfriendly==4.18
idna==2.8
importlib-resources==5.4.0
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isodate==0.6.0
jedi==0.18.0
Jinja2==2.11.2
jmespath==0.9.4
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras==2.6.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
mistune==0.8.4
mlxtend==0.19.0
monotonic==1.5
msgpack==0.6.2
multidict==5.2.0
multiprocess==0.70.12.2
mypy-extensions==0.4.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
netifaces==0.10.4
networkx==2.4
nltk==3.6.5
notebook==6.2.0
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.4.58
opt-einsum==3.3.0
packaging==20.8
palettable==3.3.0
pandas==1.2.1
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.8.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.14
protobuf==3.18.0
prov==1.5.2
psutil==5.5.1
ptyprocess==0.7.0
pyarrow==6.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.6.0
pycparser==2.20
pycups==1.9.73
Pygments==2.7.4
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
pyrsistent==0.17.3
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.0.0
qtconsole==5.0.2
QtPy==1.9.0
rdflib==4.2.2
rdflib-jsonld==0.4.0
regex==2021.10.23
requests==2.22.0
requests-aws4auth==1.0.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
roman==2.0.0
rsa==4.0
ruamel.yaml==0.15.89
s3transfer==0.3.3
schema-salad==5.0.20200220195218
scikit-learn==0.24.2
scikit-plot==0.3.7
scipy==1.6.3
screen-resolution-extra==0.0.0
seaborn==0.11.2
SecretStorage==2.3.1
Send2Trash==1.5.0
shellescape==3.4.1
simplejson==3.16.0
six==1.15.0
sklearn-pandas==2.2.0
smart-open==5.2.1
soupsieve==1.9.5
SPARQLWrapper==1.8.5
ssh-import-id==5.10
system-service==0.3
systemd-python==234
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.6.0
tensorflow-metadata==1.4.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
Theano==1.0.5
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
typing-extensions==4.0.0
ubuntu-advantage-tools==27.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.2.2
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xgboost==1.4.2
xkit==0.0.0
xxhash==2.0.2
yarl==1.7.2
zipp==3.6.0
